Feat - 191 - Extract filter accordion to a separate component and file
